requirements.txt
flask
flask-cors
onnxruntime
numpy
Pillow
